+++ dummy.go
@@ -5,8 +5,8 @@ type DummyMetrics struct{}
 func (*DummyMetrics) Alloc(_ uint32)   {}
 func (*DummyMetrics) Free(_ uint32)    {}
 func (*DummyMetrics) Release(_ uint32) {}
-func (*DummyMetrics) Set(_ uint16)     {}
-func (*DummyMetrics) Evict(_ uint16)   {}
+func (*DummyMetrics) Set(_ uint32)     {}
+func (*DummyMetrics) Evict(_ uint32)   {}
 func (*DummyMetrics) Miss()            {}
 func (*DummyMetrics) Hit()             {}
 func (*DummyMetrics) Expire()          {}
