+++ entry.go
@@ -3,7 +3,7 @@ package cbytecache
 type entry struct {
 	hash   uint64
 	offset uint32
-	length uint16
+	length uint32
 	expire uint32 // overflows at 2106-02-07 06:28:15
 	aidptr *uint32
 }
